Correcion requerimiento detalle

+++ ext-web/docroot/html/portlet/compras/requerimiento_detalle.jsp
@@ -1,6 +1,7 @@
 <%@ include file="/html/portlet/compras/init.jsp" %>
 <%@ taglib uri="http://java.sun.com/portlet_2_0" prefix="portlet" %>
-<portlet:defineObjects/>
+
+<portlet:defineObjects />
 
 <%!
 private String jsDetalleCompra(String value) {
@@ -22,10 +23,10 @@ private String jsDetalleCompraAttr(String value) {
 %>
 
 <%
-RequerimientoCompra reqDetalle = (RequerimientoCompra) renderRequest.getAttribute(WebKeysCompras.REQUERIMIENTO_COMPRA_EN_EDICION);
+RequerimientoCompra reqDetalle = (RequerimientoCompra)renderRequest.getAttribute(WebKeysCompras.REQUERIMIENTO_COMPRA_EN_EDICION);
 
 if (reqDetalle == null) {
-    reqDetalle = (RequerimientoCompra) renderRequest.getAttribute(WebKeysCompras.REQUERIMIENTO_COMPRA_EN_VIEW);
+    reqDetalle = (RequerimientoCompra)renderRequest.getAttribute(WebKeysCompras.REQUERIMIENTO_COMPRA_EN_VIEW);
 }
 
 if (reqDetalle == null) {
@@ -51,33 +52,46 @@ if (detalles == null) {
 PortletURL detalleActionURL = renderResponse.createActionURL();
 detalleActionURL.setWindowState(WindowState.MAXIMIZED);
 detalleActionURL.setParameter("struts_action", "/compras/editar_requerimiento");
+
+int detalleColspan = puedeABMDetalle ? 9 : 8;
 %>
 
 <fieldset class="block-labels">
     <legend>Detalle del requerimiento</legend>
 
     <table class="lfr-table taglib-search-iterator" width="100%">
-        <thead>
-            <tr class="portlet-section-header results-header">
-                <th>Rengl&oacute;n</th>
-                <th>Tipo art&iacute;culo</th>
-                <th>Art&iacute;culo</th>
-                <th>Cantidad</th>
-                <th>Unidad</th>
-                <th>Precio unitario estimado</th>
-                <th>Total estimado</th>
-                <th>Observaciones</th>
-                <c:if test="<%= puedeABMDetalle %>">
-                    <th>Acciones</th>
-                </c:if>
+        <tr class="portlet-section-header results-header">
+            <th>Rengl&oacute;n</th>
+            <th>Tipo art&iacute;culo</th>
+            <th>Art&iacute;culo</th>
+            <th>Cantidad</th>
+            <th>Unidad</th>
+            <th>Precio unitario estimado</th>
+            <th>Total estimado</th>
+            <th>Observaciones</th>
+
+            <% if (puedeABMDetalle) { %>
+                <th>Acciones</th>
+            <% } %>
+        </tr>
+
+        <%
+        if (detalles.size() == 0) {
+        %>
+            <tr class="portlet-section-body results-row">
+                <td colspan="<%= detalleColspan %>">
+                    No hay detalles cargados.
+                </td>
             </tr>
-        </thead>
-        <tbody>
-            <%
+        <%
+        }
+        else {
             for (int i = 0; i < detalles.size(); i++) {
                 RequerimientoCompraDetalle detalle = detalles.get(i);
-            %>
-                <tr class="<%= (i % 2 == 0) ? "portlet-section-body results-row" : "portlet-section-alternate results-row alt" %>">
+
+                String rowClass = (i % 2 == 0) ? "portlet-section-body results-row" : "portlet-section-alternate results-row alt";
+        %>
+                <tr class="<%= rowClass %>">
                     <td><%= HtmlUtil.escape(detalle.getRenglonString()) %></td>
                     <td><%= HtmlUtil.escape(detalle.getTipoArticuloVisible()) %></td>
                     <td><%= HtmlUtil.escape(detalle.getArticuloVisible()) %></td>
@@ -87,34 +101,39 @@ detalleActionURL.setParameter("struts_action", "/compras/editar_requerimiento");
                     <td><%= HtmlUtil.escape(detalle.getPrecioTotalEstimadoString()) %></td>
                     <td><%= HtmlUtil.escape(detalle.getObservacionesVisible()) %></td>
 
-                    <c:if test="<%= puedeABMDetalle %>">
+                    <% if (puedeABMDetalle) { %>
                         <td>
-                            <input type="button"
-                                   value="Editar"
-                                   onClick="<portlet:namespace />editarDetalle('<%= detalle.getIdRequerimientoDetalle() %>', '<%= jsDetalleCompraAttr(detalle.getRenglonString()) %>', '<%= jsDetalleCompraAttr(detalle.getTipoArticuloVisible()) %>', '<%= jsDetalleCompraAttr(detalle.getArticuloVisible()) %>', '<%= jsDetalleCompraAttr(detalle.getCantidadString()) %>', '<%= jsDetalleCompraAttr(detalle.getUnidadMedidaVisible()) %>', '<%= jsDetalleCompraAttr(detalle.getPrecioUnitarioEstimadoString()) %>', '<%= jsDetalleCompraAttr(detalle.getPrecioTotalEstimadoString()) %>', '<%= jsDetalleCompraAttr(detalle.getObservacionesVisible()) %>');" />
+                            <input
+                                type="button"
+                                value="Editar"
+                                onClick="<portlet:namespace />editarDetalle('<%= detalle.getIdRequerimientoDetalle() %>', '<%= jsDetalleCompraAttr(detalle.getTipoArticuloVisible()) %>', '<%= jsDetalleCompraAttr(detalle.getArticuloVisible()) %>', '<%= jsDetalleCompraAttr(detalle.getCantidadString()) %>', '<%= jsDetalleCompraAttr(detalle.getUnidadMedidaVisible()) %>', '<%= jsDetalleCompraAttr(detalle.getPrecioUnitarioEstimadoString()) %>', '<%= jsDetalleCompraAttr(detalle.getPrecioTotalEstimadoString()) %>', '<%= jsDetalleCompraAttr(detalle.getObservacionesVisible()) %>');"
+                            />
+
                             &nbsp;
-                            <form action="<%= detalleActionURL.toString() %>" method="post" name="<portlet:namespace />deleteDetalleFm<%= detalle.getIdRequerimientoDetalle() %>">
+
+                            <form action="<%= detalleActionURL.toString() %>" method="post" name="<portlet:namespace />deleteDetalleFm<%= detalle.getIdRequerimientoDetalle() %>" style="display:inline;">
                                 <input type="hidden" name="<portlet:namespace /><%= Constants.CMD %>" value="deleteItem" />
                                 <input type="hidden" name="<portlet:namespace />id_requerimiento_compra" value="<%= reqDetalle.getIdRequerimientoCompra() %>" />
                                 <input type="hidden" name="<portlet:namespace />id_requerimiento_detalle" value="<%= detalle.getIdRequerimientoDetalle() %>" />
-                                <input type="button" value="Borrar" onClick="if (confirm('&iquest;Confirma borrar el rengl&oacute;n?')) submitForm(document.<portlet:namespace />deleteDetalleFm<%= detalle.getIdRequerimientoDetalle() %>);" />
+
+                                <input
+                                    type="button"
+                                    value="Borrar"
+                                    onClick="if (confirm('&iquest;Confirma borrar el detalle?')) submitForm(document.<portlet:namespace />deleteDetalleFm<%= detalle.getIdRequerimientoDetalle() %>);"
+                                />
                             </form>
                         </td>
-                    </c:if>
+                    <% } %>
                 </tr>
-            <%
+        <%
             }
-            %>
-
-            <c:if test="<%= detalles.size() == 0 %>">
-                <tr>
-                    <td colspan="<%= puedeABMDetalle ? "9" : "8" %>">No hay renglones cargados.</td>
-                </tr>
-            </c:if>
-        </tbody>
+        }
+        %>
     </table>
 
-    <c:if test="<%= puedeABMDetalle %>">
+    <% if (puedeABMDetalle) { %>
+        <br />
+
         <form action="<%= detalleActionURL.toString() %>" method="post" name="<portlet:namespace />detalleFm">
             <input type="hidden" name="<portlet:namespace /><%= Constants.CMD %>" id="<portlet:namespace />detalle_cmd" value="addItem" />
             <input type="hidden" name="<portlet:namespace />id_requerimiento_detalle" id="<portlet:namespace />id_requerimiento_detalle" value="0" />
@@ -122,73 +141,128 @@ detalleActionURL.setParameter("struts_action", "/compras/editar_requerimiento");
 
             <table class="lfr-table" width="100%">
                 <tr>
-                    <td><label>Rengl&oacute;n:</label></td>
+                    <td><label for="<portlet:namespace />tipo_articulo">Tipo art&iacute;culo:</label></td>
                     <td>
-                        <input type="text" name="<portlet:namespace />renglon" id="<portlet:namespace />renglon" size="5" maxlength="5" />
+                        <input
+                            type="text"
+                            name="<portlet:namespace />tipo_articulo"
+                            id="<portlet:namespace />tipo_articulo"
+                            size="25"
+                            maxlength="80"
+                        />
                     </td>
 
-                    <td><label>Tipo art&iacute;culo:</label></td>
+                    <td><label for="<portlet:namespace />articulo">Art&iacute;culo:</label></td>
                     <td>
-                        <input type="text" name="<portlet:namespace />tipo_articulo" id="<portlet:namespace />tipo_articulo" size="25" maxlength="80" />
-                    </td>
-
-                    <td><label>Art&iacute;culo:</label></td>
-                    <td>
-                        <input type="text" name="<portlet:namespace />articulo" id="<portlet:namespace />articulo" size="45" maxlength="255" />
-                    </td>
-                </tr>
-
-                <tr>
-                    <td colspan="6">&nbsp;</td>
-                </tr>
-
-                <tr>
-                    <td><label>Cantidad:</label></td>
-                    <td>
-                        <input type="text" name="<portlet:namespace />cantidad" id="<portlet:namespace />cantidad" size="8" value="1" />
-                    </td>
-
-                    <td><label>Unidad:</label></td>
-                    <td>
-                        <input type="text" name="<portlet:namespace />unidad_medida" id="<portlet:namespace />unidad_medida" size="12" maxlength="30" />
-                    </td>
-
-                    <td><label>Precio unitario:</label></td>
-                    <td>
-                        <input type="text" name="<portlet:namespace />precio_unitario_estimado" id="<portlet:namespace />precio_unitario_estimado" size="12" value="0" />
+                        <input
+                            type="text"
+                            name="<portlet:namespace />articulo"
+                            id="<portlet:namespace />articulo"
+                            size="45"
+                            maxlength="255"
+                        />
                     </td>
                 </tr>
 
                 <tr>
-                    <td colspan="6">&nbsp;</td>
+                    <td colspan="4">&nbsp;</td>
                 </tr>
 
                 <tr>
-                    <td><label>Total estimado:</label></td>
+                    <td><label for="<portlet:namespace />cantidad">Cantidad:</label></td>
                     <td>
-                        <input type="text" name="<portlet:namespace />precio_total_estimado" id="<portlet:namespace />precio_total_estimado" size="12" />
+                        <input
+                            type="text"
+                            name="<portlet:namespace />cantidad"
+                            id="<portlet:namespace />cantidad"
+                            size="8"
+                            value="1"
+                        />
                     </td>
 
-                    <td><label>Observaciones:</label></td>
+                    <td><label for="<portlet:namespace />unidad_medida">Unidad:</label></td>
+                    <td>
+                        <input
+                            type="text"
+                            name="<portlet:namespace />unidad_medida"
+                            id="<portlet:namespace />unidad_medida"
+                            size="12"
+                            maxlength="30"
+                        />
+                    </td>
+                </tr>
+
+                <tr>
+                    <td colspan="4">&nbsp;</td>
+                </tr>
+
+                <tr>
+                    <td><label for="<portlet:namespace />precio_unitario_estimado">Precio unitario:</label></td>
+                    <td>
+                        <input
+                            type="text"
+                            name="<portlet:namespace />precio_unitario_estimado"
+                            id="<portlet:namespace />precio_unitario_estimado"
+                            size="12"
+                            value="0"
+                        />
+                    </td>
+
+                    <td><label for="<portlet:namespace />precio_total_estimado">Total estimado:</label></td>
+                    <td>
+                        <input
+                            type="text"
+                            name="<portlet:namespace />precio_total_estimado"
+                            id="<portlet:namespace />precio_total_estimado"
+                            size="12"
+                        />
+                    </td>
+                </tr>
+
+                <tr>
+                    <td colspan="4">&nbsp;</td>
+                </tr>
+
+                <tr>
+                    <td><label for="<portlet:namespace />observaciones_detalle">Observaciones:</label></td>
                     <td colspan="3">
-                        <input type="text" name="<portlet:namespace />observaciones_detalle" id="<portlet:namespace />observaciones_detalle" size="80" maxlength="500" />
+                        <input
+                            type="text"
+                            name="<portlet:namespace />observaciones_detalle"
+                            id="<portlet:namespace />observaciones_detalle"
+                            size="80"
+                            maxlength="500"
+                        />
                     </td>
                 </tr>
 
                 <tr>
-                    <td colspan="6">&nbsp;</td>
+                    <td colspan="4">&nbsp;</td>
                 </tr>
 
                 <tr>
-                    <td colspan="6" align="center">
-                        <input type="button" id="<portlet:namespace />detalle_submit" value="Agregar rengl&oacute;n" onClick="<portlet:namespace />agregarDetalle();" />
+                    <td colspan="4" align="center">
+                        <input
+                            type="button"
+                            id="<portlet:namespace />detalle_submit"
+                            value="Agregar detalle"
+                            onClick="<portlet:namespace />guardarDetalle();"
+                        />
+
                         &nbsp;&nbsp;
-                        <input type="button" id="<portlet:namespace />detalle_cancelar" value="Cancelar edici&oacute;n" style="display:none;" onClick="<portlet:namespace />cancelarEdicionDetalle();" />
+
+                        <input
+                            type="button"
+                            id="<portlet:namespace />detalle_cancelar"
+                            value="Cancelar edici&oacute;n"
+                            style="display:none;"
+                            onClick="<portlet:namespace />cancelarEdicionDetalle();"
+                        />
                     </td>
                 </tr>
             </table>
         </form>
-    </c:if>
+    <% } %>
 </fieldset>
 
 <script type="text/javascript">
@@ -196,10 +270,9 @@ detalleActionURL.setParameter("struts_action", "/compras/editar_requerimiento");
         return value == null ? '' : value;
     }
 
-    function <portlet:namespace />editarDetalle(idDetalle, renglon, tipoArticulo, articulo, cantidad, unidadMedida, precioUnitario, precioTotal, observaciones) {
+    function <portlet:namespace />editarDetalle(idDetalle, tipoArticulo, articulo, cantidad, unidadMedida, precioUnitario, precioTotal, observaciones) {
         jQuery('#<portlet:namespace />detalle_cmd').val('updateItem');
         jQuery('#<portlet:namespace />id_requerimiento_detalle').val(<portlet:namespace />detalleValue(idDetalle));
-        jQuery('#<portlet:namespace />renglon').val(<portlet:namespace />detalleValue(renglon));
         jQuery('#<portlet:namespace />tipo_articulo').val(<portlet:namespace />detalleValue(tipoArticulo));
         jQuery('#<portlet:namespace />articulo').val(<portlet:namespace />detalleValue(articulo));
         jQuery('#<portlet:namespace />cantidad').val(<portlet:namespace />detalleValue(cantidad));
@@ -207,7 +280,8 @@ detalleActionURL.setParameter("struts_action", "/compras/editar_requerimiento");
         jQuery('#<portlet:namespace />precio_unitario_estimado').val(<portlet:namespace />detalleValue(precioUnitario));
         jQuery('#<portlet:namespace />precio_total_estimado').val(<portlet:namespace />detalleValue(precioTotal));
         jQuery('#<portlet:namespace />observaciones_detalle').val(<portlet:namespace />detalleValue(observaciones));
-        jQuery('#<portlet:namespace />detalle_submit').val('Guardar rengl&oacute;n');
+
+        jQuery('#<portlet:namespace />detalle_submit').val('Guardar detalle');
         jQuery('#<portlet:namespace />detalle_cancelar').show();
         jQuery('#<portlet:namespace />articulo').focus();
     }
@@ -215,7 +289,6 @@ detalleActionURL.setParameter("struts_action", "/compras/editar_requerimiento");
     function <portlet:namespace />cancelarEdicionDetalle() {
         jQuery('#<portlet:namespace />detalle_cmd').val('addItem');
         jQuery('#<portlet:namespace />id_requerimiento_detalle').val('0');
-        jQuery('#<portlet:namespace />renglon').val('');
         jQuery('#<portlet:namespace />tipo_articulo').val('');
         jQuery('#<portlet:namespace />articulo').val('');
         jQuery('#<portlet:namespace />cantidad').val('1');
@@ -223,18 +296,22 @@ detalleActionURL.setParameter("struts_action", "/compras/editar_requerimiento");
         jQuery('#<portlet:namespace />precio_unitario_estimado').val('0');
         jQuery('#<portlet:namespace />precio_total_estimado').val('');
         jQuery('#<portlet:namespace />observaciones_detalle').val('');
-        jQuery('#<portlet:namespace />detalle_submit').val('Agregar rengl&oacute;n');
+
+        jQuery('#<portlet:namespace />detalle_submit').val('Agregar detalle');
         jQuery('#<portlet:namespace />detalle_cancelar').hide();
     }
 
-    function <portlet:namespace />agregarDetalle() {
-        if (jQuery('#<portlet:namespace />articulo').val() == '') {
+    function <portlet:namespace />guardarDetalle() {
+        var articulo = jQuery.trim(jQuery('#<portlet:namespace />articulo').val());
+        var cantidad = jQuery.trim(jQuery('#<portlet:namespace />cantidad').val());
+
+        if (articulo == '') {
             alert('Debe informar art&iacute;culo.');
             jQuery('#<portlet:namespace />articulo').focus();
             return;
         }
 
-        if (jQuery('#<portlet:namespace />cantidad').val() == '') {
+        if (cantidad == '') {
             alert('Debe informar cantidad.');
             jQuery('#<portlet:namespace />cantidad').focus();
             return;
